Finish backoffice admin

+++ Ease/src/com/Ease/API/V1/Admin/ServletDeleteTeam.java
@@ -2,14 +2,14 @@ package com.Ease.API.V1.Admin;
 
 import com.Ease.Hibernate.HibernateQuery;
 import com.Ease.NewDashboard.Profile;
-import com.Ease.Team.Channel;
 import com.Ease.Team.Team;
-import com.Ease.Team.TeamUser;
-import com.Ease.Utils.HttpServletException;
-import com.Ease.Utils.HttpStatus;
 import com.Ease.Utils.Servlets.PostServletManager;
+import com.Ease.websocketV1.WebSocketMessageAction;
+import com.Ease.websocketV1.WebSocketMessageFactory;
+import com.Ease.websocketV1.WebSocketMessageType;
 import com.stripe.model.Customer;
 import com.stripe.model.Subscription;
+import org.json.simple.JSONObject;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -27,62 +27,43 @@ public class ServletDeleteTeam extends HttpServlet {
         try {
             sm.needToBeEaseAdmin();
             Integer team_id = sm.getIntParam("team_id", true, false);
-            Team team;
-            try {
-                HibernateQuery hibernateQuery = sm.getHibernateQuery();
-                team = sm.getTeam(team_id);
-                for (TeamUser teamUser : team.getTeamUsers().values()) {
-                    teamUser.setAdmin_id(null);
-                    sm.saveOrUpdate(teamUser);
-                }
-                Subscription subscription = team.getSubscription();
+            Team team = sm.getTeam(team_id);
+            Subscription subscription = team.getSubscription();
+            if (subscription != null) {
                 subscription.cancel(new HashMap<>());
                 Customer customer = team.getCustomer();
                 String default_source = customer.getDefaultSource();
                 if (default_source != null && !default_source.equals(""))
                     customer.getSources().retrieve(default_source).delete();
-                team.getTeamCardMap().values().stream().flatMap(teamCard -> teamCard.getTeamCardReceiverMap().values().stream()).forEach(teamCardReceiver -> {
-                    Profile profile = teamCardReceiver.getApp().getProfile();
-                    if (profile != null) {
-                        profile.removeAppAndUpdatePositions(teamCardReceiver.getApp(), hibernateQuery);
-                        teamCardReceiver.getApp().setProfile(null);
-                        teamCardReceiver.getApp().setPosition(null);
-                    }
-                });
-                sm.deleteObject(team);
-                sm.setSuccess("Team deleted");
-            } catch (HttpServletException e) {
-                HibernateQuery hibernateQuery = sm.getHibernateQuery();
-                hibernateQuery.querySQLString("SELECT * FROM teams WHERE id = ?");
-                hibernateQuery.setParameter(1, team_id);
-                if (hibernateQuery.getSingleResult() == null)
-                    throw new HttpServletException(HttpStatus.BadRequest, "No team with this id");
-                hibernateQuery.querySQLString("UPDATE teamUsers SET admin_id = null WHERE team_id = ?");
-                hibernateQuery.setParameter(1, team_id);
-                hibernateQuery.executeUpdate();
-                hibernateQuery.queryString("SELECT t FROM Team t WHERE t.id = ?");
-                hibernateQuery.setParameter(1, team_id);
-                team = (Team) hibernateQuery.getSingleResult();
-                team.getTeamUsers().forEach((integer, teamUser) -> {
-                    teamUser.setAdmin_id(null);
-                    sm.saveOrUpdate(teamUser);
-                });
-                team.getTeamCardMap().clear();
-                team.setSubscription_id(null);
-                team.setCard_entered(false);
-                team.setActive(false);
-                sm.saveOrUpdate(team);
-                for (TeamUser teamUser : team.getTeamUsers().values())
-                    sm.deleteObject(teamUser);
-                for (Channel channel : team.getChannels().values())
-                    sm.deleteObject(channel);
-                sm.deleteObject(team);
-                sm.setSuccess("team deleted");
             }
+            HibernateQuery hibernateQuery = sm.getHibernateQuery();
+            team.getTeamCardMap().values().stream().flatMap(teamCard -> teamCard.getTeamCardReceiverMap().values().stream()).forEach(teamCardReceiver -> {
+                Profile profile = teamCardReceiver.getApp().getProfile();
+                if (profile != null) {
+                    profile.removeAppAndUpdatePositions(teamCardReceiver.getApp(), hibernateQuery);
+                    teamCardReceiver.getApp().setProfile(null);
+                    teamCardReceiver.getApp().setPosition(null);
+                }
+            });
+            team.getTeamCardMap().clear();
+            team.setSubscription_id(null);
+            team.setCard_entered(false);
+            team.setActive(false);
+            sm.setTeam(team);
+            sm.saveOrUpdate(team);
+            JSONObject ws_obj = new JSONObject();
+            ws_obj.put("team_id", team_id);
+            sm.addWebSocketMessage(WebSocketMessageFactory.createWebSocketMessage(WebSocketMessageType.TEAM, WebSocketMessageAction.REMOVED, ws_obj));
+            sm.setSuccess("Subscription ended");
         } catch (Exception e) {
+            e.printStackTrace();
             sm.setError(e);
         }
-        sm.sendResponse();
+        try {
+            sm.sendResponse();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
